Render Markdown
Added remark and remark-html packages to render file content.Changed getPostData() function to async in order to wait for content data to be processed into html, and returned this contentHtml. Changed getStaticProps() to async function too in order to wait for data to be returned. Set innerHTML to show blog content from file.